mvp funcionando

+++ rust/src/text/pages/papers.rs
@@ -1,8 +1,8 @@
 #[derive(Debug)]
 pub struct PaperProps<'a> {
     pub name: &'a str,
-    pub width: f32,
-    pub height: f32,
+    pub width: u32,
+    pub height: u32,
 }
 
-pub static A4 : PaperProps = PaperProps { name: "A4", width: 595.0, height: 842.0 };
+pub static A4 : PaperProps = PaperProps { name: "A4", width: 595, height: 842 };
